spec alignment: literal typing rules + closure cast fix
- Spec: fix auto-wrap contradiction (42 = number = f64, not i32)
- Spec: add Literal overflow section [NOT YET IMPLEMENTED]
- Closure cast: use declared param types instead of inferType(arg)
  Fixes UB: (int32_t (*)(double)) → (int32_t (*)(int32_t))
- closureParamTypes stored on sym at define-time for all closure paths
- Array-of-closures: _arrElemClosureParams for expression calls
- Rewrite 7 unknown tests: bare literal 42 + typeof i32 was never true
- New tests: narrow-f64-from-bare, narrow-f64-mismatch,
  bare-literal-number, as-cast-from-f64-unknown
- 1349 no-gcc, 1303 gcc (was 1286)

+++ test/cases/phase1/function-values/function-pointer-c/expected.c
@@ -10,6 +10,6 @@ int main(void) {
     TSC_INIT();
     tsc_closure _fns_lit[] = {(tsc_closure){.env = NULL, .fn = (void*)square_i32}};
     const Array_tsc_closure fns = {.data = _fns_lit, .length = 1, .capacity = 1};
-    printf("%d\n", ((int32_t (*)(double))fns.data[0].fn)(4));
+    printf("%d\n", ((int32_t (*)(int32_t))fns.data[0].fn)(4));
     return 0;
 }
